Changed the Flask forms to have only one input field, and to be more obviously an alliteration detection engine. Currently, the form takes in input and outputs a table, while printing the found alliteration to CL. This works as a proof of concept.
Changed the older views and forms files to be old.py files, as they are test files and no longer dependencies.
Changed the HTML files to be more aesthetic and to be *slightly* more readable.
Changed the flask forms implementation to work directly with the HTML forms and a WTForms class instead of an
unnecessary custom class. :